# contracts/openapi.v0.1.yaml
openapi: 3.1.0
info:
  title: DSH 工业仿真智能体 工程 API
  version: 0.1.2
  description: |
    dsh-sim R0 工程能力真相入口（定义书 §工程API：业务操作 / §工程API：审查、节点与上传）。
    MCP 是本 API 的受限投影。长操作返回 202 与资源 ID；请求返回不是计算结束。
    所有产生副作用的 POST 必须携带 Idempotency-Key 头（主体+项目+action+key 唯一；
    同 key 同摘要返回原对象，同 key 异摘要 → 409）。
    错误对象统一为 Error{code,message,retryable,trace_id,details}。
    本文件中的示例值均为占位：假哈希一律写作 "sha256:<64hex>" 形式，不含真实工程参数。
servers:
  - url: http://127.0.0.1:8600/api/v1
...
paths:
...
  /tasks:
    post:
      tags: [tasks]
      operationId: createTask
      summary: 创建任务草稿（符合 task-draft.schema.json）
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [draft]
              properties:
                draft:
                  type: object
                  description: 符合 https://dsh-sim.local/contracts/task-draft.schema.json
            example:
              draft:
                purpose: design_screening
                open_questions:
                  - field: conditions
                    responsible: "<engineer_id>"
                    question: "<missing input question>"
      responses:
        '201':
          description: 已创建任务
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/revisions:
    post:
      tags: [tasks]
      operationId: createRevision
      summary: 创建新修订；expected_revision 防并发覆盖
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [expected_revision, spec]
              properties:
                expected_revision: { type: integer, minimum: 0 }
                spec:
                  type: object
                  description: 符合 https://dsh-sim.local/contracts/task-spec.schema.json
      responses:
        '201':
          description: 已创建修订（发布后不可原位修改）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TaskRevision' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/prepare:
    post:
      tags: [preparations]
      operationId: prepareTask
      summary: 异步创建准备作业（PREPARE）；不在该动作开始正式求解
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision]
              properties:
                revision: { type: integer, minimum: 1 }
      responses:
        '202':
          description: 准备作业已受理，返回准备 ID
          content:
            application/json:
              schema:
                type: object
                required: [preparation_id, job_id]
                properties:
                  preparation_id: { type: string }
                  job_id: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /preparations/{preparation_id}:
    get:
      tags: [preparations]
      operationId: getPreparation
      summary: 读取准备结果：实际设置回读与差异
      parameters:
        - name: preparation_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 准备对象
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Preparation' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/authorizations:
    post:
      tags: [tasks]
      operationId: authorizeRuns
      summary: 人工运行授权；仅可信人工会话，Agent Bearer 始终拒绝
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [revision, preparation_id, prepared_digest, execution_budget, confirmation_id]
              properties:
                revision: { type: integer, minimum: 1 }
                preparation_id: { type: string }
                prepared_digest:
                  type: string
                  description: 人工确认时面板展示的 prepared 摘要
                  example: "sha256:<64hex>"
                execution_budget:
                  type: object
                  description: 预算，同 task-spec execution_budget 结构
                confirmation_id:
                  type: string
                  description: createHumanConfirmation 签发的一次性确认凭据 ID
      responses:
        '201':
          description: 授权已建立
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Authorization' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /tasks/{task_id}/submissions:
    post:
      tags: [runs]
      operationId: submitRuns
      summary: 依据已存在授权持久入队；准备摘要变化 409 拒绝
      parameters:
        - $ref: '#/components/parameters/TaskId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorization_id, prepared_digest]
              properties:
                authorization_id: { type: string }
                prepared_digest:
                  type: string
                  example: "sha256:<64hex>"
      responses:
        '202':
          description: 已受理，返回 Run 集合
          content:
            application/json:
              schema:
                type: object
                required: [run_ids]
                properties:
                  run_ids:
                    type: array
                    items: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /runs/{run_id}:
    get:
      tags: [runs]
      operationId: getRun
      summary: 查询多维状态与事件；事件按 after_seq 增量读取
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/AfterSeq'
      responses:
        '200':
          description: Run 对象（含 after_seq 之后的增量事件）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /runs/{run_id}/cancel:
    post:
      tags: [runs]
      operationId: cancelRun
      summary: 受控取消（异步请求）；取消请求不等于已停止
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '202':
          description: 取消请求已入库；全部受控子进程退出后才显示 CANCELLED
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /runs/{run_id}/retry:
    post:
      tags: [runs]
      operationId: retryRun
      summary: 补算生成新 attempt；校验原因、策略与剩余预算
      parameters:
        - name: run_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason, recovery_strategy]
              properties:
                reason: { type: string }
                recovery_strategy:
                  type: string
                  description: 已批准的恢复策略标识
      responses:
        '202':
          description: 新 attempt 已受理
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Attempt' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: 幂等键；主体+项目+action+key 唯一。同 key 同请求摘要返回原对象；同 key 异摘要 → 409。
...
    AfterSeq:
      name: after_seq
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
      description: 事件增量读取：只返回 event_seq 大于该值的事件。
    TaskId:
      name: task_id
      in: path
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: 401 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 403 越权
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 409 摘要/修订/幂等冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unprocessable:
      description: 422 工程输入不成立
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: 429 配额限制
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ServiceUnavailable:
      description: 503 暂不可用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message, retryable, trace_id, details]
      properties:
        code:
          type: string
          description: 字符串枚举错误码
          example: "<ERROR_CODE>"
        message: { type: string }
        retryable:
          type: boolean
          description: 客户端只能对 retryable 与已定义状态采取有界重试
        trace_id: { type: string, format: uuid }
        details: { type: object }
    Sha256:
      type: string
      description: SHA-256 摘要（64 位小写十六进制）；占位写作 sha256:<64hex>。哈希仅由服务计算。
      example: "sha256:<64hex>"
...
    TaskFlowState:
      enum: [DRAFT, PREPARING, READY, AUTHORIZED, ACTIVE, READY_FOR_REVIEW, IN_REVIEW, ACCEPTED, CHANGES_REQUESTED, REJECTED]
    ExecutionState:
      enum: [QUEUED, WAITING_RESOURCE, LEASED, STARTING, RUNNING, CANCELLING, COLLECTING, SUCCEEDED, FAILED, CANCELLED, LOST]
    NumericalState:
      enum: [NOT_CHECKED, PASS, FAIL, INSUFFICIENT]
    ApplicabilityState:
      enum: [IN_SCOPE, OUT_OF_SCOPE, UNCONFIRMED]
    ReviewState:
      enum: [NOT_SUBMITTED, PENDING, CHANGES_REQUESTED, ACCEPTED, REJECTED]
    ValidityState:
      enum: [CURRENT, STALE]
    Task:
      type: object
      required: [task_id, project_id, current_revision, task_state, review_state, owner_id, purpose, created_at]
      properties:
        task_id: { type: string, description: 稳定任务 ID }
        project_id: { type: string }
        current_revision: { type: integer, minimum: 0 }
        task_state: { $ref: '#/components/schemas/TaskFlowState' }
        review_state: { $ref: '#/components/schemas/ReviewState' }
        owner_id: { type: string }
        purpose: { type: string, enum: [design_screening] }
        blockers:
          type: array
          items: { type: object }
          description: 阻塞不是静默跳过；缺输入/回读失败在此表达
        created_at: { type: string, format: date-time }
    TaskRevision:
      type: object
      required: [task_id, revision, spec, spec_sha256, source_refs, created_by, created_at]
      properties:
        task_id: { type: string }
        revision: { type: integer, minimum: 1, description: 单调递增；发布后不可原位修改 }
        spec:
          type: object
          description: 完整业务输入，符合 task-spec.schema.json
        spec_sha256: { $ref: '#/components/schemas/Sha256' }
        source_refs:
          type: array
          items: { type: string }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
    Preparation:
      type: object
      required: [preparation_id, task_id, revision, prepared_digest, prepared_artifacts, readback_sha256, adapter_build, software_build, differences, blockers, created_at]
      properties:
        preparation_id: { type: string }
        task_id: { type: string }
        revision: { type: integer, minimum: 1 }
        prepared_digest: { $ref: '#/components/schemas/Sha256' }
        prepared_artifacts:
          type: object
          additionalProperties: { $ref: '#/components/schemas/Sha256' }
          description: "{logical_path: artifact_sha256}，逐文件摘要"
        readback_sha256: { $ref: '#/components/schemas/Sha256' }
        adapter_build: { type: string }
        software_build: { type: string }
        differences:
          type: array
          items: { type: object }
          description: 申请值/真实回读差异
        blockers:
          type: array
          items: { type: object }
        created_at: { type: string, format: date-time }
      description: 授权后不可变。
    Authorization:
      type: object
      required: [authorization_id, task_id, revision, preparation_id, prepared_digest, execution_budget, authorized_by, purpose, validity, created_at]
      properties:
        authorization_id: { type: string }
        task_id: { type: string }
        revision: { type: integer, minimum: 1 }
        preparation_id: { type: string }
        prepared_digest: { $ref: '#/components/schemas/Sha256' }
        execution_budget: { type: object }
        authorized_by: { type: string, description: 授权人（人工会话身份） }
        purpose: { type: string }
        validity: { $ref: '#/components/schemas/ValidityState' }
        created_at: { type: string, format: date-time }
        revoked_at: { type: string, format: date-time, nullable: true }
    Run:
      type: object
      required: [run_id, task_id, revision, variant_id, condition_id, execution_state, numerical_state, applicability_state, created_at]
      properties:
        run_id: { type: string }
        task_id: { type: string }
        revision: { type: integer, minimum: 1 }
        variant_id: { type: string }
        condition_id: { type: string }
        execution_state: { $ref: '#/components/schemas/ExecutionState' }
        numerical_state: { $ref: '#/components/schemas/NumericalState' }
        applicability_state: { $ref: '#/components/schemas/ApplicabilityState' }
        current_attempt_id: { type: string, nullable: true }
        events:
          type: array
          items: { $ref: '#/components/schemas/Event' }
          description: 按 after_seq 增量返回
        created_at: { type: string, format: date-time }
    Attempt:
      type: object
      required: [attempt_id, run_id, attempt_no, state, created_at]
      properties:
        attempt_id: { type: string, description: 不可覆盖 }
        run_id: { type: string }
        attempt_no: { type: integer, minimum: 1 }
        node_id: { type: string, nullable: true }
        lease_id: { type: string, nullable: true }
        fencing_token: { type: integer, nullable: true }
        process_identity:
          type: object
          nullable: true
          description: 进程身份/创建时间或外部调度 ID
        state: { $ref: '#/components/schemas/ExecutionState' }
        started_at: { type: string, format: date-time, nullable: true }
        ended_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
    Event:
      type: object
      required: [event_id, job_id, event_seq, kind, payload, occurred_at]
      properties:
        event_id: { type: string }
        job_id: { type: string }
        event_seq: { type: integer, minimum: 1, description: "UNIQUE(job_id,event_seq)" }
        kind:
          enum: [STARTING, RUNNING, HEARTBEAT, COMPLETED, FAILED, CANCELLED]
        payload: { type: object }
        occurred_at: { type: string, format: date-time }
...